# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The Tower API (v1)",
        "description": "Endpoints for various resources in The Tower.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
...
        "/guardians/{id}": {
            "get": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Get a guardian",
                "description": "Retrieve a single guardian.",
                "operationId": "047e192be5a0801ec0558d078ad0fc46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guardian detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuardianResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Delete guardian",
                "description": "Soft-delete a guardian.",
                "operationId": "fdd0247f282a4a336c0de2ccc6817399",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Update guardian",
                "description": "Update an existing guardian.",
                "operationId": "50672f79a16f72257e4c987ba6225d10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGuardianRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated guardian",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuardianResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/guardians": {
            "get": {
                "tags": [
                    "Guardians"
                ],
                "summary": "List guardians",
                "description": "List guardians (paginated).",
                "operationId": "68beb576f44c06a01f1846ab6d1d5b32",
                "responses": {
                    "200": {
                        "description": "Paginated list of guardians",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuardianCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Create guardian",
                "description": "Create a new guardian.",
                "operationId": "8b24d41ba56246e2374a71fdbb584834",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreGuardianRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuardianResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/guardians/{id}/restore": {
            "patch": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Restore guardian",
                "description": "Restore a soft-deleted guardian.",
                "operationId": "196f13d001343cfcf4da12b842a478c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restored guardian",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuardianResource"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Guardian is not deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreGuardianRequest": {
                "description": "Payload sent when creating a new Guardian",
                "required": [
                    "name",
                    "value"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "Epic Battle Theme"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.006
                    }
                },
                "type": "object"
            },
...
            "UpdateGuardianRequest": {
                "description": "Payload for updating an existing Guardian. All fields optional.",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "Epic Battle Theme Remix"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.007
                    }
                },
                "type": "object"
            },
...
            "GuardianCollection": {
                "description": "Paginated list of Guardians",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GuardianResource"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/guardians?page=1"
                            },
                            "last": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/guardians?page=2"
                            },
                            "prev": {
                                "type": "string",
                                "format": "uri",
                                "example": null,
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/guardians?page=2",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 2
                            },
                            "path": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/guardians"
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 50
                            },
                            "to": {
                                "type": "integer",
                                "example": 50
                            },
                            "total": {
                                "type": "integer",
                                "example": 100
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "GuardianResource": {
                "description": "Represents a Guardian available in the API",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Muse"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.006
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}